Moving the GitHub-specific exceptions into their own subpackage
