+++ src/main/java/usi/si/seart/exception/github/GitHubAPIException.java
@@ -1,4 +1,4 @@
-package usi.si.seart.exception;
+package usi.si.seart.exception.github;
 
 import lombok.experimental.StandardException;
 import usi.si.seart.github.ErrorResponse;
