+++ src/main/java/usi/si/seart/github/GitHubAPIConnector.java
@@ -28,7 +28,7 @@ import org.springframework.retry.support.RetryTemplate;
 import org.springframework.stereotype.Component;
 import org.springframework.web.client.HttpClientErrorException;
 import org.springframework.web.client.HttpServerErrorException;
-import usi.si.seart.exception.GitHubAPIException;
+import usi.si.seart.exception.github.GitHubAPIException;
 import usi.si.seart.util.Ranges;
 
 import java.net.URL;
